Fix libfmt build errors (#828)
* Fix build errors with libfmt 10.2.1

* Use `fmt::print` instead of `std::cout`

(cherry picked from commit e5e935060622fcd15e6e80dad877662f2587571f)

+++ clients/common/containers/d_vector.hpp
@@ -72,8 +72,8 @@ public:
             auto status = (hipMemset)(d, 0, bytes);
             if(status != hipSuccess)
             {
-                fmt::print(stderr, "error: {} ({}) at {}:{}\n", hipGetErrorString(status), status,
-                           __FILE__, __LINE__);
+                fmt::print(stderr, "error: {} ({}) at {}:{}\n", hipGetErrorString(status),
+                           static_cast<std::int32_t>(status), __FILE__, __LINE__);
                 rocblas_abort();
             }
         }
